requirements.txt
aioconsole==0.6.2
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
MouseInfo==0.1.3
multidict==6.0.4
mysql-connector-python==8.2.0
Pillow==10.1.0
pregex==2.3.3
protobuf==4.21.12
psutil==5.9.6
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.0.7
yarl==1.9.2
